requirements.txt
Flask>=3.1,<4
flask-cors>=6,<7
gunicorn>=23,<24

redis>=5,<8
requests>=2.32,<3
python-dotenv>=1.0,<2

APScheduler>=3.11,<4